+++ pages/reports/transactions.vue
@@ -3,49 +3,31 @@ import VueApexCharts from 'vue3-apexcharts';
 import api from '~/lib/api';
 import { CHART_COLORS } from '~/lib/consts';
 import { useAuth } from '~/composables/use_auth';
-import { parseNumberArray, parseStringArray } from '~/lib/helper_parsers';
+import { useChart } from '~/composables/use_chart';
 
-const { token } = useAuth();
 const route = useRoute();
-const router = useRouter();
-const appConfig = useAppConfig();
+const { token } = useAuth();
+const {
+  CHART_HEIGTH,
+  PERIODS,
+  period,
+  setPeriod,
+  filters,
+  onCategoryClick,
+  onAccountClick,
+  onProjectClick,
+  onPropertyClick,
+  onKindClick,
+  labelColor,
+} = useChart();
 
+const CHART_TYPE = 'bar';
 const isLoading = ref(true);
 const isError = ref(false);
 const chartData = ref({});
 
-const CHART_HEIGTH = 500;
-const CHART_TYPE = 'bar';
-const currentMonthLabel = new Intl.DateTimeFormat('ru-RU', {
-  month: 'long',
-}).format(new Date());
-const labelColor = appConfig.theme.dark ? '#e2e8f0' : '#334155';
-
-const PERIODS = computed(() => ({
-  CURRENT_MONTH: currentMonthLabel.charAt(0).toUpperCase() + currentMonthLabel.slice(1),
-  DAYS_30: '30 дней',
-  DAYS_60: '60 дней',
-  YEARS_1: 'Год',
-  YEARS_2: 'Два года',
-  YEARS_5: 'Пять лет',
-  ALL: 'Всё время',
-}));
-const isPeriodValid = (value) => Boolean(PERIODS.value[value]);
-const period = ref(isPeriodValid(route.query.period) ? route.query.period : 'CURRENT_MONTH');
-
 const series = computed(() => chartData.value.series);
 const categories = computed(() => chartData.value.categories);
-
-const filters = computed(() => {
-  return {
-    period: period.value,
-    accountIds: parseNumberArray(route.query.accounts),
-    categoryIds: parseNumberArray(route.query.categories),
-    projectIds: parseNumberArray(route.query.projects),
-    propertyIds: parseNumberArray(route.query.properties),
-    kinds: parseStringArray(route.query.kinds),
-  };
-});
 
 const selectedAccounts = ref([]);
 const selectedCategories = ref([]);
@@ -100,56 +82,6 @@ const load = async () => {
   }
 };
 
-const setPeriod = (value) => {
-  period.value = value;
-
-  router.push({
-    query: {
-      ...route.query,
-      period: value,
-    },
-  });
-};
-
-const toggleQueryFilter = (queryKey, id, parser, formatter = (arr) => arr.join(',')) => {
-  const current = parser(route.query[queryKey]);
-  const newValues = current.includes(id)
-    ? current.filter(item => item !== id)
-    : [...current, id];
-  
-  const nextQuery = { ...route.query };
-  if (newValues.length) nextQuery[queryKey] = formatter(newValues);
-  else delete nextQuery[queryKey];
-  
-  router.replace({ query: nextQuery });
-};
-
-const onCategoryClick = (id) => {
-  toggleQueryFilter('categories', id, parseNumberArray);
-};
-
-const onAccountClick = (id) => {
-  toggleQueryFilter('accounts', id, parseNumberArray);
-};
-
-const onProjectClick = (id) => {
-  toggleQueryFilter('projects', id, parseNumberArray);
-};
-
-const onPropertyClick = (id) => {
-  toggleQueryFilter('properties', id, parseNumberArray);
-};
-
-const onKindClick = (id) => {
-  toggleQueryFilter('kinds', id, parseStringArray);
-};
-
-watch(() => route.query.period, (newPeriod) => {
-  if (isPeriodValid(newPeriod) && newPeriod !== period.value) {
-    period.value = newPeriod;
-  }
-});
-
 watch(
   () => route.query,
   () => {
@@ -195,7 +127,6 @@ const chartOptions = computed(() => ({
       }
     },
     tooltip: { enabled: false },
-    // type: 'datetime',
   },
   yaxis: {
     labels: {
@@ -262,7 +193,7 @@ const chartOptions = computed(() => ({
               v-for='kind in selectedKinds'
               :key='kind.id'
               :name='kind.name'
-              :is-x='true'              
+              :is-x='true'
               @click='onKindClick(kind.id)'
             />
